chore: add boundary test
